chore(data): ⚰️ move old data to legacy
