+++ src/data/legacy/README.md
@@ -1,0 +1,31 @@
+# 📚 Project Documentation
+
+## 🌐 Overview
+
+This document outlines key components and files within our project repository to ensure clarity and understanding for developers, contributors, and stakeholders.
+
+## 📁 FeatureProjects.ts File
+
+### 🚦 Current Status
+
+The `FeatureProjects.ts` file, which previously contained important project data, is currently not in active use. The decision to cease its use came as a result of our migration to Supabase for data management. Supabase now serves as the primary source for our project's data needs, offering enhanced scalability, security, and ease of access compared to the static data previously held within `FeatureProjects.ts`.
+
+### 🤔 Rationale for Retention
+
+Despite its current dormancy, the `FeatureProjects.ts` file remains in the repository for several critical reasons:
+
+- **🔬 Future Testing and Experimentation:** The file may provide a useful framework for testing or developing features in isolation from the main Supabase data store. Its structure and example data can be instrumental in exploring new ideas or conducting experiments without impacting live data.
+
+- **📖 Historical Context:** Retaining `FeatureProjects.ts` offers valuable insights into the project's development history. It helps document the evolution of data management approaches within the project, from static files to dynamic, cloud-based solutions.
+
+- **🚀 Quick Prototyping:** For rapid prototyping or when working offline, `FeatureProjects.ts` offers a readily available data set. This can be particularly useful for new developers getting up to speed or when testing functionalities that do not require the latest data.
+
+### 📝 Guidance for Developers
+
+Developers are encouraged to be aware of the `FeatureProjects.ts` file and its potential uses, especially in contexts where offline development or isolated testing is beneficial. However, it's important to remember that any new development or data integration should prioritize Supabase as the main data source to ensure consistency and integrity across the project.
+
+Should there be a need to reference or reactivate `FeatureProjects.ts` for specific scenarios, developers are advised to discuss this with the project's lead developers or the data management team to align on the best approach.
+
+## ℹ️ Further Information
+
+For additional details on Supabase integration, other project components, or specific documentation requests, please consult the table of contents or reach out directly to the project management team.
